MOB-1874 Support EXIF orientation setting to rotate bitmaps before display and upload
Temporary File is created for each file shared via the extension. The bitmap is rotated (if needed) at that time.
In the app (compose message activity, documents activity), we already work with Files so we don't need to create them temporarily, we just get the orientation setting and rotate the bitmap if needed.