MOB-1912 Set click listeners in code rather than layout
On Android 4.0.x, the click listeners defined in the onClick attribute of the component is not detected by the activity that is parent of the fragment.
Therefore clicking on the components did not trigger anything. To fix that, the click listeners are now defined in code when the fragment is created.