MOB-1932 Add actions on drive folders
Refactor ExoFile to include its children files and folders.
Use these in the methods that deal with content.
Use the root ExoFile of the requested content to get all information necessary to display or not the action button.
Create ExoDocumentUtils.getFileFromXMLElement() and use it in other methods.
Remove/Add some comments.